my-python-app/requirements.txt
streamlit
pymupdf
numpy
python-dotenv
sentence-transformers
langchain
google-generativeai
sarvamai